Factor GCN Circular front matter into a separate component
We're going to reuse this component in tool tips.
